hashmap_from_file should not be part of mijia crate.
